Convert uint16_t to int for WebRTC cipher/crypto suite.
This is a follow up CL on https://codereview.webrtc.org/1337673002

BUG=
R=pthatcher@webrtc.org

Review URL: https://codereview.webrtc.org/1377733004 .

Cr-Commit-Position: refs/heads/master@{#10175}

+++ talk/app/webrtc/statscollector.cc
@@ -734,7 +734,7 @@ void StatsCollector::ExtractSessionInfo() {
         channel_report->AddString(StatsReport::kStatsValueNameSrtpCipher,
                                   srtp_cipher);
       }
-      uint16_t ssl_cipher = channel_iter.ssl_cipher;
+      int ssl_cipher = channel_iter.ssl_cipher;
       if (ssl_cipher &&
           rtc::SSLStreamAdapter::GetSslCipherSuiteName(ssl_cipher).length()) {
         channel_report->AddString(
